move Collection contract
