+++ src/Support/DataStructure/Arrayee.php
@@ -2,7 +2,7 @@
 
 namespace ArtARTs36\MergeRequestLinter\Support\DataStructure;
 
-use ArtARTs36\MergeRequestLinter\Contracts\Collection;
+use ArtARTs36\MergeRequestLinter\Contracts\DataStructure\Collection;
 use ArtARTs36\MergeRequestLinter\Support\DataStructure\Traits\ContainsAll;
 use ArtARTs36\MergeRequestLinter\Support\DataStructure\Traits\CountProxy;
 use Traversable;
