+++ src/Support/DataStructure/Set.php
@@ -2,7 +2,7 @@
 
 namespace ArtARTs36\MergeRequestLinter\Support\DataStructure;
 
-use ArtARTs36\MergeRequestLinter\Contracts\Collection;
+use ArtARTs36\MergeRequestLinter\Contracts\DataStructure\Collection;
 use ArtARTs36\MergeRequestLinter\Support\ArrayKeyIterator;
 use ArtARTs36\MergeRequestLinter\Support\DataStructure\Traits\CountProxy;
