+++ tests/Mocks/MockEvaluatingSubject.php
@@ -2,8 +2,8 @@
 
 namespace ArtARTs36\MergeRequestLinter\Tests\Mocks;
 
-use ArtARTs36\MergeRequestLinter\Contracts\Collection;
 use ArtARTs36\MergeRequestLinter\Contracts\Condition\EvaluatingSubject;
+use ArtARTs36\MergeRequestLinter\Contracts\DataStructure\Collection;
 use ArtARTs36\MergeRequestLinter\Support\DataStructure\Arrayee;
 
 class MockEvaluatingSubject implements EvaluatingSubject
